+++ pennylane_lightning/core/src/measurements/tests/Test_MeasurementsBase.cpp
@@ -19,6 +19,8 @@ namespace {
 using Pennylane::Util::isApproxEqual;
 } // namespace
 /// @endcond
+#include <algorithm>
+#include <string>
 
 #ifdef _ENABLE_PLQUBIT
 constexpr bool BACKEND_FOUND = true;
@@ -144,6 +146,161 @@ template <typename TypeList> void testProbabilities() {
 TEST_CASE("Probabilities", "[MeasurementsBase]") {
     if constexpr (BACKEND_FOUND) {
         testProbabilities<TestStateVectorBackends>();
+    }
+}
+
+template <typename TypeList> void testProbabilitiesObs() {
+    if constexpr (!std::is_same_v<TypeList, void>) {
+        using StateVectorT = typename TypeList::Type;
+        using PrecisionT = typename StateVectorT::PrecisionT;
+
+        const size_t num_qubits = 3;
+
+        // Defining the Statevector that will be measured.
+        auto statevector_data = createNonTrivialState<StateVectorT>();
+        StateVectorT statevector(statevector_data.data(),
+                                 statevector_data.size());
+
+        StateVectorT sv(statevector_data.data(), statevector_data.size());
+
+        DYNAMIC_SECTION("Test PauliX"
+                        << StateVectorToName<StateVectorT>::name) {
+            for (size_t i = 0; i < num_qubits; i++) {
+                NamedObs<StateVectorT> obs("PauliX", {i});
+                Measurements<StateVectorT> Measurer_shots(statevector);
+
+                sv.applyOperation("Hadamard", {i}, false);
+
+                Measurements<StateVectorT> Measurer(sv);
+
+                auto prob_shots = Measurer_shots.probs(obs);
+                auto prob = Measurer.probs({i});
+
+                REQUIRE_THAT(prob_shots, Catch::Approx(prob).margin(1e-6));
+            }
+        }
+
+        DYNAMIC_SECTION("Test PauliY"
+                        << StateVectorToName<StateVectorT>::name) {
+            for (size_t i = 0; i < num_qubits; i++) {
+                NamedObs<StateVectorT> obs("PauliY", {i});
+                Measurements<StateVectorT> Measurer_shots(statevector);
+
+                sv.applyOperations({"PauliZ", "S", "Hadamard"}, {{i}, {i}, {i}},
+                                   {false, false, false});
+
+                Measurements<StateVectorT> Measurer(sv);
+
+                auto prob_shots = Measurer_shots.probs(obs);
+                auto prob = Measurer.probs({i});
+
+                REQUIRE_THAT(prob_shots, Catch::Approx(prob).margin(1e-6));
+            }
+        }
+
+        DYNAMIC_SECTION("Test PauliZ"
+                        << StateVectorToName<StateVectorT>::name) {
+            for (size_t i = 0; i < num_qubits; i++) {
+                NamedObs<StateVectorT> obs("PauliZ", {i});
+                Measurements<StateVectorT> Measurer_shots(statevector);
+
+                Measurements<StateVectorT> Measurer(sv);
+
+                auto prob_shots = Measurer_shots.probs(obs);
+                auto prob = Measurer.probs({i});
+
+                REQUIRE_THAT(prob_shots, Catch::Approx(prob).margin(1e-6));
+            }
+        }
+
+        DYNAMIC_SECTION("Test Hadamard"
+                        << StateVectorToName<StateVectorT>::name) {
+            for (size_t i = 0; i < num_qubits; i++) {
+                NamedObs<StateVectorT> obs("Hadamard", {i});
+                Measurements<StateVectorT> Measurer_shots(statevector);
+                const PrecisionT theta = -M_PI / 4.0;
+                sv.applyOperation("RY", {i}, false, {theta});
+
+                Measurements<StateVectorT> Measurer(sv);
+
+                auto prob_shots = Measurer_shots.probs(obs);
+                auto prob = Measurer.probs({i});
+
+                REQUIRE_THAT(prob_shots, Catch::Approx(prob).margin(1e-6));
+            }
+        }
+
+        DYNAMIC_SECTION("Test Identity"
+                        << StateVectorToName<StateVectorT>::name) {
+            for (size_t i = 0; i < num_qubits; i++) {
+                NamedObs<StateVectorT> obs("Identity", {i});
+                Measurements<StateVectorT> Measurer_shots(statevector);
+
+                Measurements<StateVectorT> Measurer(sv);
+
+                auto prob_shots = Measurer_shots.probs(obs);
+                auto prob = Measurer.probs({i});
+
+                REQUIRE_THAT(prob_shots, Catch::Approx(prob).margin(1e-6));
+            }
+        }
+
+        DYNAMIC_SECTION("Test TensorProd XYZ"
+                        << StateVectorToName<StateVectorT>::name) {
+            auto X0 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliX", std::vector<size_t>{0});
+            auto Z1 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliZ", std::vector<size_t>{1});
+            auto Y2 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliY", std::vector<size_t>{2});
+            auto obs = TensorProdObs<StateVectorT>::create({X0, Z1, Y2});
+
+            Measurements<StateVectorT> Measurer_shots(statevector);
+
+            sv.applyOperations({"Hadamard", "PauliZ", "S", "Hadamard"},
+                               {{0}, {2}, {2}, {2}},
+                               {false, false, false, false});
+
+            Measurements<StateVectorT> Measurer(sv);
+
+            auto prob_shots = Measurer_shots.probs(*obs);
+            auto prob = Measurer.probs({0, 1, 2});
+
+            REQUIRE_THAT(prob_shots, Catch::Approx(prob).margin(1e-6));
+        }
+
+        DYNAMIC_SECTION("Test TensorProd YHI"
+                        << StateVectorToName<StateVectorT>::name) {
+            auto Y0 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliY", std::vector<size_t>{0});
+            auto H1 = std::make_shared<NamedObs<StateVectorT>>(
+                "Hadamard", std::vector<size_t>{1});
+            auto I2 = std::make_shared<NamedObs<StateVectorT>>(
+                "Identity", std::vector<size_t>{2});
+            auto obs = TensorProdObs<StateVectorT>::create({Y0, H1, I2});
+
+            Measurements<StateVectorT> Measurer_shots(statevector);
+
+            sv.applyOperations({"PauliZ", "S", "Hadamard"}, {{0}, {0}, {0}},
+                               {false, false, false});
+            const PrecisionT theta = -M_PI / 4.0;
+            sv.applyOperation("RY", {1}, false, {theta});
+
+            Measurements<StateVectorT> Measurer(sv);
+
+            auto prob_shots = Measurer_shots.probs(*obs);
+            auto prob = Measurer.probs({0, 1, 2});
+
+            REQUIRE_THAT(prob_shots, Catch::Approx(prob).margin(1e-6));
+        }
+
+        testProbabilitiesObs<typename TypeList::Next>();
+    }
+}
+
+TEST_CASE("Probabilities Obs shots", "[MeasurementsBase]") {
+    if constexpr (BACKEND_FOUND) {
+        testProbabilitiesObs<TestStateVectorBackends>();
     }
 }
 
@@ -511,6 +668,20 @@ template <typename TypeList> void testNamedObsVar() {
                     REQUIRE(expected == Approx(result).margin(1e-6));
                 }
             }
+
+            DYNAMIC_SECTION(obs_name[ind_obs]
+                            << " Shots - Varying wires"
+                            << StateVectorToName<StateVectorT>::name) {
+                for (size_t ind_wires = 0; ind_wires < wires_list.size();
+                     ind_wires++) {
+                    NamedObs<StateVectorT> obs(obs_name[ind_obs],
+                                               wires_list[ind_wires]);
+                    PrecisionT expected = exp_values_ref[ind_obs][ind_wires];
+                    size_t num_shots = 10000;
+                    PrecisionT result = Measurer.var(obs, num_shots);
+                    REQUIRE(expected == Approx(result).margin(5e-2));
+                }
+            }
         }
         testNamedObsVar<typename TypeList::Next>();
     }
@@ -598,6 +769,58 @@ TEST_CASE("Var - HermitianObs", "[MeasurementsBase][Observables]") {
     }
 }
 
+template <typename TypeList> void testTensorProdObsVarShot() {
+    if constexpr (!std::is_same_v<TypeList, void>) {
+        using StateVectorT = typename TypeList::Type;
+        using PrecisionT = typename StateVectorT::PrecisionT;
+        using ComplexT = StateVectorT::ComplexT;
+
+        // Defining the State Vector that will be measured.
+        std::vector<ComplexT> statevector_data{
+            {0.0, 0.0}, {0.0, 0.1}, {0.1, 0.1}, {0.1, 0.2},
+            {0.2, 0.2}, {0.3, 0.3}, {0.3, 0.4}, {0.4, 0.5}};
+        StateVectorT statevector(statevector_data.data(),
+                                 statevector_data.size());
+
+        // Initializing the measures class.
+        // This object attaches to the statevector allowing several measures.
+        Measurements<StateVectorT> Measurer(statevector);
+
+        DYNAMIC_SECTION(" Without Identity"
+                        << StateVectorToName<StateVectorT>::name) {
+            size_t num_shots = 10000;
+            auto X0 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliX", std::vector<size_t>{0});
+            auto Z1 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliZ", std::vector<size_t>{1});
+            auto obs = TensorProdObs<StateVectorT>::create({X0, Z1});
+            auto expected = Measurer.var(*obs);
+            auto result = Measurer.var(*obs, num_shots);
+            REQUIRE(expected == Approx(result).margin(5e-2));
+        }
+
+        DYNAMIC_SECTION(" With Identity"
+                        << StateVectorToName<StateVectorT>::name) {
+            size_t num_shots = 10000;
+            auto X0 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliX", std::vector<size_t>{0});
+            auto I1 = std::make_shared<NamedObs<StateVectorT>>(
+                "Identity", std::vector<size_t>{1});
+            auto obs = TensorProdObs<StateVectorT>::create({X0, I1});
+            PrecisionT expected = Measurer.var(*obs);
+            PrecisionT result = Measurer.var(*obs, num_shots);
+            REQUIRE(expected == Approx(result).margin(5e-2));
+        }
+
+        testTensorProdObsVarShot<typename TypeList::Next>();
+    }
+}
+
+TEST_CASE("Var Shot- TensorProdObs", "[MeasurementsBase][Observables]") {
+    if constexpr (BACKEND_FOUND) {
+        testTensorProdObsVarShot<TestStateVectorBackends>();
+    }
+}
 template <typename TypeList> void testSamples() {
     if constexpr (!std::is_same_v<TypeList, void>) {
         using StateVectorT = typename TypeList::Type;
@@ -664,6 +887,160 @@ TEST_CASE("Samples", "[MeasurementsBase]") {
     }
 }
 
+template <typename TypeList> void testSamplesCountsObs() {
+    if constexpr (!std::is_same_v<TypeList, void>) {
+        using StateVectorT = typename TypeList::Type;
+        using PrecisionT = typename StateVectorT::PrecisionT;
+
+        // Defining the State Vector that will be measured.
+        auto statevector_data = createNonTrivialState<StateVectorT>();
+        StateVectorT statevector(statevector_data.data(),
+                                 statevector_data.size());
+
+        constexpr size_t twos[] = {
+            1U << 0U,  1U << 1U,  1U << 2U,  1U << 3U,  1U << 4U,  1U << 5U,
+            1U << 6U,  1U << 7U,  1U << 8U,  1U << 9U,  1U << 10U, 1U << 11U,
+            1U << 12U, 1U << 13U, 1U << 14U, 1U << 15U, 1U << 16U, 1U << 17U,
+            1U << 18U, 1U << 19U, 1U << 20U, 1U << 21U, 1U << 22U, 1U << 23U,
+            1U << 24U, 1U << 25U, 1U << 26U, 1U << 27U, 1U << 28U, 1U << 29U,
+            1U << 30U, 1U << 31U};
+
+        // Initializing the measures class.
+        // This object attaches to the statevector allowing several measures.
+        Measurements<StateVectorT> Measurer(statevector);
+
+        std::vector<std::vector<size_t>> wires_list = {{0}, {1}, {2}};
+        std::vector<std::string> obs_name = {"PauliX", "PauliY", "PauliZ",
+                                             "Hadamard", "Identity"};
+        // Expected results calculated with Pennylane default.qubit:
+        std::vector<std::vector<PrecisionT>> exp_values_ref = {
+            {0.49272486, 0.42073549, 0.28232124},
+            {-0.64421768, -0.47942553, -0.29552020},
+            {0.58498357, 0.77015115, 0.91266780},
+            {0.7620549436, 0.8420840225, 0.8449848566},
+            {1.0, 1.0, 1.0}};
+        for (size_t ind_obs = 0; ind_obs < obs_name.size(); ind_obs++) {
+            DYNAMIC_SECTION(obs_name[ind_obs]
+                            << " Sample Obs - Varying wires"
+                            << StateVectorToName<StateVectorT>::name) {
+                size_t num_shots = 10000;
+                for (size_t ind_wires = 0; ind_wires < wires_list.size();
+                     ind_wires++) {
+                    NamedObs<StateVectorT> obs(obs_name[ind_obs],
+                                               wires_list[ind_wires]);
+                    PrecisionT expected = exp_values_ref[ind_obs][ind_wires];
+                    auto samples = Measurer.sample(obs, num_shots);
+
+                    PrecisionT result = 0.0;
+                    for (auto &it : samples) {
+                        result += it;
+                    }
+                    result /= num_shots;
+
+                    REQUIRE(expected == Approx(result).margin(5e-2));
+                }
+            }
+
+            DYNAMIC_SECTION(obs_name[ind_obs]
+                            << " Counts Obs - Varying wires"
+                            << StateVectorToName<StateVectorT>::name) {
+                size_t num_shots = 10000;
+                for (size_t ind_wires = 0; ind_wires < wires_list.size();
+                     ind_wires++) {
+                    NamedObs<StateVectorT> obs(obs_name[ind_obs],
+                                               wires_list[ind_wires]);
+                    PrecisionT expected = exp_values_ref[ind_obs][ind_wires];
+                    auto samples = Measurer.counts(obs, num_shots);
+
+                    PrecisionT result = 0.0;
+                    for (auto &it : samples) {
+                        result += it.first * it.second;
+                    }
+                    result /= num_shots;
+
+                    REQUIRE(expected == Approx(result).margin(5e-2));
+                }
+            }
+        }
+
+        DYNAMIC_SECTION("samples() without obs"
+                        << StateVectorToName<StateVectorT>::name) {
+            std::vector<PrecisionT> expected_probabilities = {
+                0.67078706, 0.03062806, 0.0870997,  0.00397696,
+                0.17564072, 0.00801973, 0.02280642, 0.00104134};
+
+            size_t num_qubits = 3;
+            size_t N = std::pow(2, num_qubits);
+            size_t num_samples = 100000;
+            auto &&samples = Measurer.sample(num_samples);
+
+            std::vector<size_t> counts(N, 0);
+            std::vector<size_t> samples_decimal(num_samples, 0);
+
+            // convert samples to decimal and then bin them in counts
+            for (size_t i = 0; i < num_samples; i++) {
+                for (size_t j = 0; j < num_qubits; j++) {
+                    if (samples[i * num_qubits + j] != 0) {
+                        samples_decimal[i] += twos[(num_qubits - 1 - j)];
+                    }
+                }
+                counts[samples_decimal[i]] += 1;
+            }
+
+            // compute estimated probabilities from histogram
+            std::vector<PrecisionT> probabilities(counts.size());
+            for (size_t i = 0; i < counts.size(); i++) {
+                probabilities[i] = counts[i] / (PrecisionT)num_samples;
+            }
+
+            REQUIRE_THAT(probabilities,
+                         Catch::Approx(expected_probabilities).margin(.05));
+        }
+
+        DYNAMIC_SECTION("counts() without obs"
+                        << StateVectorToName<StateVectorT>::name) {
+            std::vector<std::string> expected_keys = {
+                "000", "001", "010", "011", "100", "101", "110", "111"};
+
+            std::vector<PrecisionT> expected_probabilities = {
+                0.67078706, 0.03062806, 0.0870997,  0.00397696,
+                0.17564072, 0.00801973, 0.02280642, 0.00104134};
+
+            size_t num_qubits = 3;
+            size_t N = std::pow(2, num_qubits);
+            size_t num_samples = 100000;
+
+            auto &&counts_sample = Measurer.counts(num_samples);
+
+            std::vector<size_t> counts(N, 0);
+
+            // convert samples to decimal and then bin them in counts
+            for (auto &it : counts_sample) {
+                auto key = it.first;
+
+                counts[key] = it.second;
+            }
+
+            // compute estimated probabilities from histogram
+            std::vector<PrecisionT> probabilities(counts.size());
+            for (size_t i = 0; i < counts.size(); i++) {
+                probabilities[i] = counts[i] / (PrecisionT)num_samples;
+            }
+
+            REQUIRE_THAT(probabilities,
+                         Catch::Approx(expected_probabilities).margin(.05));
+        }
+
+        testSamplesCountsObs<typename TypeList::Next>();
+    }
+}
+
+TEST_CASE("Samples Obs", "[MeasurementsBase]") {
+    if constexpr (BACKEND_FOUND) {
+        testSamplesCountsObs<TestStateVectorBackends>();
+    }
+}
+
 template <typename TypeList> void testHamiltonianObsExpvalShot() {
     if constexpr (!std::is_same_v<TypeList, void>) {
         using StateVectorT = typename TypeList::Type;
@@ -718,6 +1095,59 @@ template <typename TypeList> void testHamiltonianObsExpvalShot() {
 TEST_CASE("Expval Shot - HamiltonianObs ", "[MeasurementsBase][Observables]") {
     if constexpr (BACKEND_FOUND) {
         testHamiltonianObsExpvalShot<TestStateVectorBackends>();
+    }
+}
+
+template <typename TypeList> void testHamiltonianObsVarShot() {
+    if constexpr (!std::is_same_v<TypeList, void>) {
+        using StateVectorT = typename TypeList::Type;
+
+        // Defining the State Vector that will be measured.
+        auto statevector_data = createNonTrivialState<StateVectorT>();
+        StateVectorT statevector(statevector_data.data(),
+                                 statevector_data.size());
+
+        // Initializing the measures class.
+        // This object attaches to the statevector allowing several measures.
+        Measurements<StateVectorT> Measurer(statevector);
+
+        DYNAMIC_SECTION("YZ" << StateVectorToName<StateVectorT>::name) {
+            auto Y0 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliY", std::vector<size_t>{0});
+            auto Z1 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliZ", std::vector<size_t>{1});
+
+            auto ob = Hamiltonian<StateVectorT>::create({0.5, 0.5}, {Y0, Z1});
+
+            size_t num_shots = 10000;
+
+            auto res = Measurer.var(*ob, num_shots);
+            auto expected = Measurer.var(*ob);
+            REQUIRE(expected == Approx(res).margin(5e-2));
+        }
+
+        DYNAMIC_SECTION("YI" << StateVectorToName<StateVectorT>::name) {
+            auto Y0 = std::make_shared<NamedObs<StateVectorT>>(
+                "PauliY", std::vector<size_t>{0});
+            auto I1 = std::make_shared<NamedObs<StateVectorT>>(
+                "Identity", std::vector<size_t>{1});
+
+            auto ob = Hamiltonian<StateVectorT>::create({0.5, 0.5}, {Y0, I1});
+
+            size_t num_shots = 10000;
+
+            auto res = Measurer.var(*ob, num_shots);
+            auto expected = Measurer.var(*ob);
+            REQUIRE(expected == Approx(res).margin(5e-2));
+        }
+
+        testHamiltonianObsVarShot<typename TypeList::Next>();
+    }
+}
+
+TEST_CASE("Var Shot - HamiltonianObs ", "[MeasurementsBase][Observables]") {
+    if constexpr (BACKEND_FOUND) {
+        testHamiltonianObsVarShot<TestStateVectorBackends>();
     }
 }