+++ pennylane_lightning/core/src/simulators/lightning_gpu/measurements/MeasurementsGPUMPI.hpp
@@ -210,6 +210,18 @@ class MeasurementsMPI final
     }
 
     /**
+     * @brief Probabilities to measure rotated basis states.
+     *
+     * @param obs An observable object.
+     *
+     * @return Floating point std::vector with probabilities
+     * in lexicographic order.
+     */
+    std::vector<PrecisionT> probs(const Observable<StateVectorT> &obs) {
+        return BaseType::probs(obs);
+    }
+
+    /**
      * @brief Utility method for samples.
      *
      * @param num_samples Number of Samples
@@ -734,5 +746,19 @@ class MeasurementsMPI final
 
         return mean_square - squared_mean * squared_mean;
     };
+
+    /**
+     * @brief Calculate the variance for an observable with the number of shots.
+     *
+     * @param obs An observable object.
+     * @param num_shots Number of shots.
+     *
+     * @return Variance of the given observable.
+     */
+
+    auto var(const Observable<StateVectorT> &obs, const size_t &num_shots)
+        -> PrecisionT {
+        return BaseType::var(obs, num_shots);
+    }
 }; // class Measurements
 } // namespace Pennylane::LightningGPU::Measures